requirements/docker.txt
appdirs==1.4.3
astroid==1.4.9
bleach==2.0.0
certifi==2017.4.17
chardet==3.0.4
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.2
django-bootstrap3==7.1.0
django-bootstrap3-datetimepicker-2==2.4.2
django-countries==4.0
django-extensions==1.7.9
django-formset-js==0.5.0
django-jquery-js==3.1.1
django-model-utils==2.6
django-recaptcha2==1.0.3
django-summernote==0.8.6
django-widget-tweaks==1.4.1
django-import-export==0.5.1
entrypoints==0.2.2
gunicorn==19.7.0
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
jedi==0.10.2
Jinja2==2.9.6
jsonfield==1.0.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.5.2
pylint==1.6.5
pyparsing==2.2.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.2
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
wrapt==1.10.10
django-mail-templated==2.6.5
django-wkhtmltopdf==3.1.0
paypalrestsdk==1.13.1
mysqlclient==1.3.12
